runner/master Store jobs in cache, & factorization
